requirements/production.txt
asgiref==3.2.3
Django==2.2.10
django-ckeditor==5.9.0
django-crispy-forms==1.9.0
django-filter==2.2.0
django-hitcount==1.3.1
django-js-asset==1.2.2
djangorestframework==3.11.0
gunicorn==20.0.4
jsmin==2.2.2
Pillow==7.0.0
python-decouple==3.3
pytz==2019.3
sqlparse==0.3.0
whitenoise==5.0.1
